fix: Update signing certificate (#47)
The public key had expired.  It's now been extended and uploaded to
keys.openpgp.org.  The new keys will expire on 2029-11-07.

As part of extending the expiry date on the public key, a new cert
needed to be generated, encrypted and uploaded to the travis server.

That is the file included in this PR.

Reference:
- https://github.com/IBM/ibm-cloud-sdk-common/blob/main/JavaDeploy.md#3-create-a-publicprivate-key-pair-for-signing-artifacts

Signed-off-by: Mike Melick <melickm@ca.ibm.com>